cmd redo split done
