+++ src/execution/exec_command.c
@@ -6,7 +6,7 @@
 /*   By: lilypad <lilypad@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/25 22:06:28 by ylang             #+#    #+#             */
-/*   Updated: 2026/02/27 17:31:18 by lilypad          ###   ########.fr       */
+/*   Updated: 2026/02/27 17:51:46 by lilypad          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,243 +24,6 @@ stateful builtin → 不能 fork
 
 stateless builtin → 可 fork 可不 fork**
 */
-
-void	free_strs(char **envp)
-{
-	int	i;
-
-	i = 0;
-	if (!envp)
-		return ;
-	while (envp[i])
-		free(envp[i++]);
-	free(envp);
-}
-
-int	count_exported_env_vars(t_list *node)
-{
-	t_env_var	*env_var;
-	int			count;
-
-	count = 0;
-	while (node)
-	{
-		env_var = env_var_from_node(node);
-		if (env_var && env_var->exported && env_var->name)
-			count++;
-		node = node->next;
-	}
-	return (count);
-}
-
-char	*build_env_entry(t_env_var *env_var)
-{
-	size_t	name_len;
-	size_t	val_len;
-	char	*val;
-	char	*entry;
-
-	if (!env_var || !env_var->name)
-		return (NULL);
-	name_len = ft_strlen(env_var->name);
-	if (env_var->value)
-		val = env_var->value;
-	else
-		val = "";
-	val_len = ft_strlen(val);
-	entry = malloc(name_len + 1 + val_len + 1);
-	if (!entry)
-		return (NULL);
-	ft_memcpy(entry, env_var->name, name_len);
-	entry[name_len] = '=';
-	ft_memcpy(entry + name_len + 1, val, val_len);
-	entry[name_len + 1 + val_len] = '\0';
-	return (entry);
-}
-
-char	**free_partial_envp(char **envp, int i)
-{
-	while (i > 0)
-		free(envp[--i]);
-	free(envp);
-	return (NULL);
-}
-
-/* 1) count exported */
-/* 2) build envp */
-/* rollback */
-char	**env_list_to_envp(t_list *env_list)
-{
-	int			i;
-	t_list		*node;
-	t_env_var	*env_var;
-	char		**envp;
-
-	i = 0;
-	node = env_list;
-	envp = malloc(sizeof(char *) * (count_exported_env_vars(node) + 1));
-	if (!envp)
-		return (NULL);
-	node = env_list;
-	while (node)
-	{
-		env_var = env_var_from_node(node);
-		if (env_var && env_var->exported && env_var->name)
-		{
-			envp[i] = build_env_entry(env_var);
-			if (!envp[i])
-				return (free_partial_envp(envp, i));
-			i++;
-		}
-		node = node->next;
-	}
-	envp[i] = NULL;
-	return (envp);
-}
-char	*search_path_value(t_list *env)
-{
-	t_env_var	*env_var;
-
-	while (env)
-	{
-		env_var = env_var_from_node(env);
-		if (!env_var)
-		{
-			env = env->next;
-			continue ;
-		}
-		if (ft_strcmp(env_var->name, "PATH") == 0)
-			return (env_var->value);
-		env = env->next;
-	}
-	return (NULL);
-}
-static char	*find_executable(char **dirs, char *cmd)
-{
-	char	*mid_path;
-	char	*full_path;
-
-	while (*dirs)
-	{
-		mid_path = ft_strjoin(*dirs, "/");
-		if (!mid_path)
-			return (NULL);
-		full_path = ft_strjoin(mid_path, cmd);
-		free(mid_path);
-		if (!full_path)
-			return (NULL);
-		if (access(full_path, X_OK) == 0)
-			return (full_path);
-		free(full_path);
-		dirs++;
-	}
-	return (NULL);
-}
-// PATH=/home/ylang/bin:/home/ylang/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin
-// try ./cmd
-// check for execution permission
-// free dirs up to now
-// free strs too
-char	*resolve_cmd_path(char *cmd, t_shell_context *sh_ctx)
-{
-	char	**dirs;
-	char	*path_ptr;
-	t_list	*env;
-	char	*result;
-
-	path_ptr = NULL;
-	env = sh_ctx->env;
-	path_ptr = search_path_value(env);
-	if (!path_ptr)
-		return (ft_strjoin("./", cmd));
-	if (*path_ptr == '\0')
-		return (NULL);
-	dirs = ft_split(path_ptr, ':');
-	if (!dirs)
-		return (NULL);
-	result = find_executable(dirs, cmd);
-	free_strs(dirs);
-	return (result);
-}
-
-int	execve_errno_to_status(int err)
-{
-	if (err == ENOENT)
-		return (127);
-	return (126);
-}
-
-int	validate_command(t_ast_command *cmd)
-{
-	if (!cmd || !cmd->args || !cmd->args[0])
-		return (0);
-	if (ft_strcmp(cmd->args[0], ".") == 0)
-		return (print_msg_n_return(127, cmd->args[0], NULL,
-				"filename argument required"));
-	if (ft_strcmp(cmd->args[0], "..") == 0)
-		return (print_msg_n_return(127, cmd->args[0], NULL,
-				"command not found"));
-	if (*cmd->args[0] == '\0')
-		return (print_msg_n_return(127, cmd->args[0], NULL,
-				"command not found"));
-	return (-1);
-}
-
-int	check_valid_path(char *path, char *cmd_name)
-{
-	struct stat	st;
-
-	if (stat(path, &st) == -1)
-	{
-		if (errno == ENOENT)
-			return (print_errno_n_return(127, path, NULL, errno));
-		if (errno == EACCES)
-			return (print_errno_n_return(126, path, NULL, errno));
-		if (errno == EISDIR)
-			return (print_errno_n_return(126, path, NULL, errno));
-		if (errno == EFAULT)
-			return (print_msg_n_return(127, cmd_name, NULL,
-					"command not found"));
-		return (print_errno_n_return(127, path, NULL, errno));
-	}
-	return (-1);
-}
-
-// cmd->arg[0]=="."
-// cmd->args[0] ==".."
-// cmd->args[0] ==""
-// contains '/', search for cmd from current directory
-// permission denied
-// the direcotry
-// no / present, so need to search and execute external via PATH,
-int	execute_external(t_ast_command *cmd, t_shell_context *sh_ctx)
-{
-	int		status;
-	char	*path;
-	char	**envp;
-
-	status = validate_command(cmd);
-	if (status != -1)
-		return (status);
-	if (ft_strchr(cmd->args[0], '/'))
-		path = cmd->args[0];
-	else
-		path = resolve_cmd_path(cmd->args[0], sh_ctx);
-	status = check_valid_path(path, cmd->args[0]);
-	if (status != -1)
-		return (status);
-	envp = env_list_to_envp(sh_ctx->env);
-	if (!envp)
-		return (print_msg_n_return(1, cmd->args[0], NULL,
-				"fail to convert env_list to envp"));
-	execve(path, cmd->args, envp);
-	status = execve_errno_to_status(errno);
-	free_strs(envp);
-	if (path != cmd->args[0])
-		free(path);
-	return (print_errno_n_return(status, cmd->args[0], NULL, errno));
-}
-
 // Fallback: should not happen in normal minishell execution
 int	wait_status_to_shell_status(int wait_status)
 {
@@ -272,8 +35,8 @@ int	wait_status_to_shell_status(int wait_status)
 		return (128 + WSTOPSIG(wait_status));
 	return (1);
 }
-// only parent / interactive shell should print
 
+// only parent / interactive shell should print
 void	report_child_termination_signal(int wait_status, const char *cmd_name,
 		t_shell_context *ctx)
 {
@@ -344,6 +107,7 @@ int	execute_builtin(t_ast_command *cmd, t_shell_context *ctx)
 		return (1);
 	return (func(cmd->args, ctx));
 }
+
 // check if it is built-in
 // if it is built-in like cd, export, unset, ecit et,
 // run directly in currentg process
